change variation id

+++ MasterOnline/Controllers/TokopediaController.cs
@@ -4311,7 +4311,7 @@ namespace MasterOnline.Controllers
 
         public class GetVariantResultData
         {
-            public long variant_id { get; set; }
+            public int variant_id { get; set; }
             public string name { get; set; }
             public string identifier { get; set; }
             public int status { get; set; }
